Move storybook stories into their respective component directories
