+++ ui/web/.storybook/config.js
@@ -11,7 +11,7 @@ const themes = {
 addDecorator(withThemes(themes));
 
 // automatically import all files ending in *.stories.js
-const req = require.context('../stories', true, /\.stories\.js$/);
+const req = require.context('../src/components', true, /\.stories\.js$/);
 function loadStories() {
   req.keys().forEach(filename => req(filename));
 }
